refactor: changing dtos extensions
